ai-layer/requirements.txt
# AURACARE/ai-layer/requirements.txt

# Core AI and Gemini
google-generativeai==0.3.2
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0

# Utilities
requests==2.31.0
python-multipart==0.0.6

# Optional: For better JSON handling
jsonschema==4.20.0

# Optional: For audio processing (if implementing audio features)
# speechrecognition==3.10.0
# pydub==0.25.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1